refactor(hooks): extract gate/meta subscriptions from useSessionHostEvents
The single subscription useEffect carried ~30 listeners inline; the tail
326 lines were one coherent domain — turn_error, permission, ask_user
(+cleared), plan bar + completion history, title, index_changed (Remote
IM disk writes).

Moved them verbatim into src/hooks/sessionHostGateSubscriptions.ts as
registerGateAndMetaSubscriptions(c, env), where env passes the epoch
machinery (isCancelled/track/listenWithRetry) owned by the parent
effect. Same single-effect lifecycle, registration ordering, and retry
semantics — the subscription list just reads as domains now.

useSessionHostEvents.ts: 2237 -> 1907 lines.
